[FTR] Mockup Images
Feature Description

Providing the images for the mockup for describing the overall application.
These are rough mockups due to lack of UI mockup tools and images but is a
starting point to get the idea of the application across.

Since these images are publicly hosted on Github it is easy to reference them
in the associated markdown documents.

The images are for non-commerical use only due to `NinjaMocks` user-agreement
for their services.

Related Issues
- #2