[FTR] Firebase Database Communication
Feature Description

Commit contains the base logic for communicating with the Firebase
database service API.  At the moment the data being saved is trival
and constant (just a string).

Modified the UI to provide a button to force sending of data to the
database to do some basic setup.

One small note is that for receiving data back from the database using
Firebase, need to have the POJO for the class cast to support setter
and getter methods to the class attributes in addition to an empty
class constructor.

Related Issues
- #9