[FTR] Record Time Fragment
Feature Description

Population of the fragment by the `MainActivity` activity class.  At
the moment the fragment manager is populating the default fragment which
has the initial layout of the stop watch into the main activity.

This shows that the UI for the stop watch is being shown and can move to
populating the fields appropriately.

Related Issues
- #21