[FTR] CountUpTimer Supporting API
Feature Description

The `CountUpTimer` object can now support API calls of `pause`, `stop`, and
`onTick` that allows for restoring previous time to get timing a bit closer to
the desired amount.  If the caller provided `onTick` amount is equal to the
future time then this means that the possible imprecision is equal to the future
time value.

Related Issues
- #21
- #22