[FTR] Timer Play Pause Stop Replay
Feature Description

Timer class now supports the play, puase, stop, replay capabilities.  It is
up to the controlling object to determine the logic as to the order to call
and teh callbacks to supply.

The default of the timer is that it will act as a count up timer based on
the future time desired along with the interval check amount.  The class
exposed API means that combinations of the play, pause, stop actions can
change the value of which the object time is operating with but not the
object itself, the class will control that.

Related Issues
- #21
- #22