[FTR] StopWatch Class
Feature Description

Providing of the `StopWatch` java object that is responsible for managing
the millisecond count that of which the `CountUpTimer` is responsible.  This
addition of a new class is to remove the counting logic from the UI Fragment.

The UI fragment just needs to convert the time to a string object and handle the
correct state for displaying the elapsed counter time.

Related Issues
- #21