[FTR] Updates Up to Second Accuracy
Feature Description

When a tick event occurs then the time value is built to display the stopwatch
activity time amount.  Previously was adding up the wrong value then rebasing.

Now just storing the previous time remaining in the class mMillisecond value
and comparing it with the time remaining.  When a `onFinish` is called then resets
the mMillisecond to the interval amount.

Related Issues
- #22