[FTR] Setting up RecyclerView, Adapater, ViewHolder
Feature Description

For the time being will use a recyclerview widget for displaying the lap
times over the `ListView`.  Might go back and change later but want to
get the feature up and working first along with the logic.

At the moment the view hold and adapter are private class members within
the stopwatch fragment.

Related Issues
- #21
- #27