[FTR] Pause & Reset State
Feature Description

Previously a pause to play transistion would keep the last time as the time
that was last seen from the tick event resolution.  This meant once play was
pressed the displayed time would include the time that elapsed within the
pause state.

Now for the transition from pause to play state in the stop watch, the
prevTime for the difference calculation is set prior to finalizing the state
transition.

Additionally, the reset had been neglected and has now been updated to perform
the same functionality as the stop but also followed by the play.

Related Issues
- #29