[FTR] StopWatchActions Fragment to Activity Interface
Feature Description

The first supported set of communication that will need to be supported
is the passing of the set of actions for the stopwatch UI back to the
actual activity to determine what to do with the overall state of the
stopwatch object.

Interface is accomplished by the StopWatchActivity defining the Actions
fragment's interface definition as a class and then it being available
upon the fragment's attachment to the fragment manager.

Related Issues
- #31