[FTR] Elapsed Time Fragment Display
Feature Description

The Elapsed Time fragment will now be capable of receiving the string representation
of the time elapsed and will now display it.  Default at the moment is based off
building the time string with a time value of 0.

The building of a StopWatch string has been moved to the StopWatch object class as a
public static member function.

Related Issues
- #31