[FTR] Skipping Time & LifeCycle Adjustments
Feature Description

In the event that life-cycle event occur, wanted to address that there is a chance
that the fragments are not setup for when the delievery of a `onTick` event from
the `StopWatch` activity.  To reduce this chance the `onPause` lifecycle event
will now cause an unregister of the callback of the `StopWatchActivity` with its
`StopWatch` object.  When the activity returns with an onCreate or resume a
register event of the callback occurs.

Additionally fixed the jumping of time due to the first value in the laps is the
current elapsed time as well.  Therefore do not need to sum all the lap times to
obtain the elapsed time.  Only need to account for the first lap time in the event
that there is more than one lap time.  This is just the string representation that
would likely get fixed on the first `onTick` but the `StopWatchActivity` state
could be stopped with invalid represnetation.

Related Issues
- #31
- #32
- #33
- #34