[FTR] Graph Adjustments with Static Data
Feature Description

Can now graph stacked column data.  The restriction at the moment is that the
lap times for each session start at the top and the last / most recent laps for
the session are near the bottom.

Coloring of the column stacks now will be correct with the top mosts being the
frist lap representataion per session of a given activity.

Additionally each lap time presented is colored.  The total number of colors will
need to equal to the session with the maximum number of laps.

Current removed all grid and axis labels except for the x-axis which will indicate
the session that it represents.

Related Issues
- #37
- #39
- #40
- #41