[FTR] onClickSupport of TimedActivity
Feature Description

Cleanup of the layout of the timed activity element that is contained within the
recyclerview widget for the collection of timed activities.

Additionally, added new interface to the `TimedActivityAdapter` to support the
onClick event of a specific timed element that will be responsible for launching the
activity's detailed view.

Related Issues
- #45
- #49