[FTR] Processing of Activity Sessions
Feature Description

Clicking on an activity will now display the activity details.  The lap
times associated with that detail will be displayed in the graph below
with bars being equal in size to the total number of seconds per lap.

At the moment, that static data generation is limited to four due to the
fact that the color palate being used is limited to four without adding
addition colors.

Related Issues
- #49