[FTR] Activity with Sessions Firebase Support
Feature Description

Current layout does support and allow to have the `ActivitySession` data
to be populated within the `TimedACtivity` Firebase entry that is
written.

Will do this for now since it is a one way relationship (e.g. a session
belongs to an activity but an activity does not belong to a session).

Related Issues
- #63