[FTR] Saving Returned StopWatch Time
Feature Description

The total time and laps that was executed by the user will now be saved
in the `ActivityTimerActivity` class scope private variables.  If the
totaltime is greater than 0 then the activity selection will occur,
otherwise the private member variables will be reset to their default
values.

Related Issues
- #63