[FTR] onChildAdded Within Firebase
Feature Description

Currently supporting just the addition of a new child.  With the current
logic the push will create a new DB entry within the database.  This is
only desired for a new activity w.r.t the current user.

For an already existing activity, will need to perform an update by
adding in the new session, and update the total time associated.

Related Issues
- #63