[FTR] Selectable Activities
Feature Description

Taking advantage of the Callback state stack.  The current callback state stack
will actually cache the activities prior to switching from the stopwatch activity
and the select activity but after the `onPause` is called.  Odd behavior and might
be limited to the device I am using.  So to take advantage of this to have a valid
selection list just provided the list of all the activities.

If it ends up being device specific, can pass in the maps as well, as just lists
and then rebuild afterward.  Would expect a null pointer exception however since
the state of teh firebase db connect to then be invalid.

Related Issues
- #62
- #64
- #65