[BUG] Enforce Time of Check to Time of Use
BUG Description

The creation of a new database reference occurred even though the authentication
listener was dropped.  This doesn't mean that the user wasn't signed in, just that
the states of the listeners to their respective references were not in a valid state.

The state the bug was in that I could have a valid db listener without a valid auth
listener which is not a supported state but did provide desired functionality.

A forced state check of the listener is now performed prior to performing an attach
of a new db listener to the db reference.

Related Issues
- #87