[FTR] SelectActivity Query
Feature Description

Supporting of feature set to allow for query of the user's collection of
timed activities in order to select one that the new timed session belongs
to.

`onResume` is responsible for the creation and attaching of the listener
to the query object.

`onPause` is responsible for the detaching and null of the query object.

Will now no longer need to pass in all the activities associated with a
user since it will be queried for directly within the activity.

Related Issues
- #89