[FTR] SelectActivity Selection
Feature Description

Selection of the activity from the list of available now has the behavior
of the adapter providing the activity the index, key, and timed activity
object that can be updated.

Unselect will change the color back to white, as will the toggle button
for a new activity usage.

Related Issues
- #89