[FTR] Class Protections & Static Private Class Cleanup
Feature Description

Made public class that is to be of static context only to have a private
constructor so it can not be instaniated.  Additionally modified the private
class within the select-activity activity that is responsible for pushing
changes to now be static private class.

This require update of the attributes to not use the SelectActivity class
members.  Additionally had to setup a static context definition to obtain
the strings associated with the APK.

Related Issues
- #102
- #104